BatchWrite: support multi tables
