+++ core/src/main/scala/com/pingcap/tispark/TiDBDataSource.scala
@@ -16,6 +16,7 @@
 package com.pingcap.tispark
 
 import com.pingcap.tikv.exception.TiBatchWriteException
+import com.pingcap.tispark.write.{TiDBOptions, TiDBWriter}
 import org.apache.spark.sql.sources._
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql._
