+++ core/src/main/scala/com/pingcap/tispark/write/EncodedKVPair.scala
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2020 PingCAP, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.pingcap.tispark.write
+
+/**
+ * EncodedKVPair is used at last stage if write process.
+ * At this stage, sorting should not be happened anymore.
+ */
+case class EncodedKVPair(encodedKey: SerializableKey, encodedValue: Array[Byte]) {
+  override def hashCode(): Int = encodedKey.hashCode()
+}
