+++ core/src/main/scala/com/pingcap/tispark/write/TiDBWriter.scala
@@ -1,6 +1,22 @@
-package com.pingcap.tispark
+/*
+ * Copyright 2020 PingCAP, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.pingcap.tispark.write
 
 import com.pingcap.tikv.exception.TiBatchWriteException
+import com.pingcap.tispark.TiDBUtils
 import org.apache.spark.sql._
 
 object TiDBWriter {
@@ -20,7 +36,7 @@ object TiDBWriter {
           if (tableExists) {
             saveMode match {
               case SaveMode.Append =>
-                TiBatchWrite.writeToTiDB(df, tiContext, options)
+                TiBatchWrite.write(df, tiContext, options)
 
               case _ =>
                 throw new TiBatchWriteException(
