+++ core/src/main/scala/com/pingcap/tispark/write/WrappedEncodedRow.scala
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2020 PingCAP, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.pingcap.tispark.write
+
+import TiBatchWrite.TiRow
+
+case class WrappedEncodedRow(row: TiRow,
+                             handle: Long,
+                             encodedKey: SerializableKey,
+                             encodedValue: Array[Byte],
+                             isIndex: Boolean,
+                             indexId: Long,
+                             remove: Boolean)
+    extends Ordered[WrappedEncodedRow] {
+  override def compare(that: WrappedEncodedRow): Int = this.handle.toInt - that.handle.toInt
+
+  override def hashCode(): Int = encodedKey.hashCode()
+}
